:recycle:2차 리팩토링 완료
