+++ client/src/pages/Join.tsx
@@ -3,9 +3,7 @@ import "../scss/join.scss";
 import JoinFormContainer from "../components/join/JoinFormContainer";
 
 const Join: React.FC = () => {
-  return (
-    <JoinFormContainer />
-  );
+  return <JoinFormContainer />;
 };
 
 export default Join;
